requirements.txt
pytest==7.3.1
jsonschema==4.17.3
requests==2.31.0
allure-pytest==2.13.2
curlify==2.2.1